+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/GeckoResult.java
@@ -6,6 +6,7 @@ import org.mozilla.gecko.util.ThreadUtils;
 import android.os.Handler;
 import android.os.Looper;
 import android.os.SystemClock;
+import android.support.annotation.AnyThread;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
@@ -151,6 +152,7 @@ import java.util.concurrent.TimeoutException;
  *
  * @param <T> The type of the value delivered via the GeckoResult.
  */
+@AnyThread
 public class GeckoResult<T> {
     private static final String LOGTAG = "GeckoResult";
 
@@ -557,6 +559,7 @@ public class GeckoResult<T> {
          * @return Result used to complete the next result in the chain. May be null.
          * @throws Throwable Exception used to complete next result in the chain.
          */
+        @AnyThread
         @Nullable GeckoResult<U> onValue(@Nullable T value) throws Throwable;
     }
 
@@ -575,6 +578,7 @@ public class GeckoResult<T> {
          * @return Result used to complete the next result in the chain. May be null.
          * @throws Throwable Exception used to complete next result in the chain.
          */
+        @AnyThread
         @Nullable GeckoResult<V> onException(@NonNull Throwable exception) throws Throwable;
     }